+++ serde_int_map_traits/src/lib.rs
@@ -5,7 +5,7 @@ pub trait UnknownKeyHandler {
 
     fn new() -> Self;
     fn num_items(&self) -> usize;
-    fn iter(&self) -> std::collections::hash_map::Iter<u32, Self::ValueType>;
-    fn handles_key(&self, key: u32) -> bool;
-    fn fill_value(&mut self, key: u32, value: Self::ValueType);
+    fn iter(&self) -> std::collections::hash_map::Iter<i64, Self::ValueType>;
+    fn handles_key(&self, key: i64) -> bool;
+    fn fill_value(&mut self, key: i64, value: Self::ValueType);
 }
